move logo into assets and update imports
